rename apiHelper file to comply with naming conventions
